requirements.txt
fastapi==0.115.12
uvicorn==0.27.1
python-dotenv==1.1.1
psycopg2-binary==2.9.10
sqlmodel==0.0.24
alembic==1.16.4
#new
PyJWT==2.10.1
python-multipart==0.0.20